requirements.txt
pyYaml
google-apis-client-generator
numpy
google-cloud-storage
google-cloud-logging